:twisted_rightwards_arrows: Merge(feature/stock Branch -> dev Branch): 사용자 하드코딩 수정

+++ src/main/java/com/heartfoilo/demo/domain/stock/api/LikeStockController.java
@@ -2,6 +2,7 @@ package com.heartfoilo.demo.domain.stock.api;
 
 import com.heartfoilo.demo.domain.stock.dto.responseDto.LikeStockResponseDto;
 import com.heartfoilo.demo.domain.stock.service.LikeService;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -17,28 +18,34 @@ public class LikeStockController {
 
 
     @PostMapping("/{stockId}")
-    public ResponseEntity<Void> addFavorite(@PathVariable("stockId") Long stockId) {
-        // TODO: 사용자 인증 추가시 변경 - OAuth2 인증 로직을 추가해야 합니다.
-        Long userId = 1L; // 임시로 하드코딩된 사용자 ID, 추후 OAuth2 인증 후 userId를 가져오는 로직으로 변경
+    public ResponseEntity<Void> addFavorite(@PathVariable("stockId") Long stockId, HttpServletRequest request) {
+        // Interceptor에서 설정한 userId 속성을 가져옴
+        Long userId = (Long) request.getAttribute("userId");
+
+        if (userId == null) {
+            return ResponseEntity.status(401).build(); // Unauthorized 처리
+        }
 
         likeService.addFavorite(userId, stockId);
         return ResponseEntity.ok().build();
     }
 
     @GetMapping
-    public ResponseEntity<List<LikeStockResponseDto>> getFavorite() {
-        // TODO: 사용자 인증 추가시 변경 - OAuth2 인증 로직을 추가해야 합니다.
-        Long userId = 1L; // 임시로 하드코딩된 사용자 ID, 추후 OAuth2 인증 후 userId를 가져오는 로직으로 변경
-
+    public ResponseEntity<List<LikeStockResponseDto>> getFavorite(HttpServletRequest request) {
+        Long userId = (Long) request.getAttribute("userId");
+        if (userId == null) {
+            return ResponseEntity.status(401).build(); // Unauthorized 처리
+        }
         List<LikeStockResponseDto> favorites = likeService.getFavorites(userId);
         return ResponseEntity.ok(favorites);
     }
 
     @DeleteMapping("/{stockId}")
-    public ResponseEntity<Void> removeFavorite(@PathVariable("stockId") Long stockId) {
-        // TODO: 사용자 인증 추가시 변경 - OAuth2 인증 로직을 추가해야 합니다.
-        Long userId = 1L; // 임시로 하드코딩된 사용자 ID, 추후 OAuth2 인증 후 userId를 가져오는 로직으로 변경
-
+    public ResponseEntity<Void> removeFavorite(@PathVariable("stockId") Long stockId, HttpServletRequest request) {
+        Long userId = (Long) request.getAttribute("userId");
+        if (userId == null) {
+            return ResponseEntity.status(401).build(); // Unauthorized 처리
+        }
         likeService.removeFavorite(userId, stockId);
         return ResponseEntity.ok().build();
     }
